style: modify header component for mobile & pc
