+++ assets/styles/index.scss
@@ -1,4 +1,5 @@
 @import "element-plus/dist/index.css";
+@import "element-plus/theme-chalk/display.css";
 
 @import "./variables";
 @import "./layouts";
