+++ components/common/Header/Mobile.vue
@@ -1,0 +1,11 @@
+<template>
+    mobile header
+</template>
+
+<script setup lang="ts">
+
+</script>
+
+<style lang="scss" scoped>
+
+</style>
